+++ console/ErrorHandler.php
@@ -42,8 +42,11 @@ class ErrorHandler extends \yii\base\ErrorHandler
             $message .= $this->formatMessage(" '" . get_class($exception) . "'", [Console::BOLD, Console::FG_BLUE])
                 . " with message " . $this->formatMessage("'{$exception->getMessage()}'", [Console::BOLD]) //. "\n"
                 . "\n\nin " . dirname($exception->getFile()) . DIRECTORY_SEPARATOR . $this->formatMessage(basename($exception->getFile()), [Console::BOLD])
-                . ':' . $this->formatMessage($exception->getLine(), [Console::BOLD, Console::FG_YELLOW]) . "\n\n"
-                . $this->formatMessage("Stack trace:\n", [Console::BOLD]) . $exception->getTraceAsString();
+                . ':' . $this->formatMessage($exception->getLine(), [Console::BOLD, Console::FG_YELLOW]) . "\n";
+            if ($exception instanceof \yii\db\Exception && !empty($exception->errorInfo)) {
+                $message .= "\n" . $this->formatMessage("Error Info:\n", [Console::BOLD]) . print_r($exception->errorInfo, true);
+            }
+            $message .= "\n" . $this->formatMessage("Stack trace:\n", [Console::BOLD]) . $exception->getTraceAsString();
         } else {
             $message = $this->formatMessage('Error: ') . $exception->getMessage();
         }