+++ helpers/HtmlPurifier.php
@@ -24,7 +24,7 @@ namespace yii\helpers;
  * ]);
  * ```
  *
- * For more details please refer to HTMLPurifier documentation](http://htmlpurifier.org/).
+ * For more details please refer to [HTMLPurifier documentation](http://htmlpurifier.org/).
  *
  * Note that you should add `ezyang/htmlpurifier` to your composer.json `require` section and run `composer install`
  * before using it.