+++ requirements/YiiRequirementChecker.php
@@ -287,7 +287,6 @@ class YiiRequirementChecker
             $minCheckResult = true;
         }
         if ($max !== null) {
-            var_dump($postMaxSize, $uploadMaxFileSize, $max);
             $maxCheckResult = $this->compareByteSize($postMaxSize, $max, '<=') && $this->compareByteSize($uploadMaxFileSize, $max, '<=');
         } else {
             $maxCheckResult = true;